+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/electric/machines/AutoEnchanter.java
@@ -38,11 +38,6 @@ public class AutoEnchanter extends AContainer {
     }
 
     @Override
-    public String getInventoryTitle() {
-        return "&5Auto-Enchanter";
-    }
-
-    @Override
     public ItemStack getProgressBar() {
         return new ItemStack(Material.GOLDEN_CHESTPLATE);
     }