+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/electric/machines/ElectricGoldPan.java
@@ -42,11 +42,6 @@ public abstract class ElectricGoldPan extends AContainer implements RecipeDispla
     }
 
     @Override
-    public String getInventoryTitle() {
-        return "&6Electric Gold Pan";
-    }
-
-    @Override
     public ItemStack getProgressBar() {
         return new ItemStack(Material.DIAMOND_SHOVEL);
     }