+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/electric/machines/ElectricPress.java
@@ -50,7 +50,7 @@ public abstract class ElectricPress extends AContainer implements RecipeDisplayI
         addRecipe(4, new ItemStack(Material.COAL, 9), new ItemStack(Material.COAL_BLOCK));
         addRecipe(4, new ItemStack(Material.SAND, 4), new ItemStack(Material.SANDSTONE));
         addRecipe(4, new ItemStack(Material.RED_SAND, 4), new ItemStack(Material.RED_SANDSTONE));
-        
+
         addRecipe(5, new ItemStack(Material.IRON_INGOT, 9), new ItemStack(Material.IRON_BLOCK));
         addRecipe(5, new ItemStack(Material.GOLD_INGOT, 9), new ItemStack(Material.GOLD_BLOCK));
 
@@ -63,12 +63,6 @@ public abstract class ElectricPress extends AContainer implements RecipeDisplayI
 
     private void addRecipe(int seconds, ItemStack input, ItemStack output) {
         registerRecipe(seconds, new ItemStack[] { input }, new ItemStack[] { output });
-    }
-
-    @Override
-    public String getInventoryTitle() {
-        return "&eElectric Press";
-
     }
 
     @Override