+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/electric/machines/ElectricSmeltery.java
@@ -44,7 +44,7 @@ public abstract class ElectricSmeltery extends AContainer {
     public ElectricSmeltery(Category category, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe) {
         super(category, item, recipeType, recipe);
 
-        new BlockMenuPreset(getID(), getInventoryTitle()) {
+        new BlockMenuPreset(getID(), getItemName()) {
 
             @Override
             public void init() {
@@ -151,11 +151,6 @@ public abstract class ElectricSmeltery extends AContainer {
                 }
             });
         }
-    }
-
-    @Override
-    public String getInventoryTitle() {
-        return "&cElectric Smeltery";
     }
 
     @Override