+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/electric/machines/ElectrifiedCrucible.java
@@ -41,9 +41,4 @@ public abstract class ElectrifiedCrucible extends AContainer {
         return new ItemStack(Material.FLINT_AND_STEEL);
     }
 
-    @Override
-    public String getInventoryTitle() {
-        return "&4Electrified Crucible";
-    }
-
 }