+++ src/main/java/io/github/thebusybiscuit/slimefun4/utils/itemstack/SlimefunGuideItem.java
@@ -1,0 +1,46 @@
+package io.github.thebusybiscuit.slimefun4.utils.itemstack;
+
+import java.util.LinkedList;
+import java.util.List;
+
+import org.bukkit.Material;
+import org.bukkit.inventory.ItemStack;
+import org.bukkit.inventory.meta.ItemMeta;
+
+import io.github.thebusybiscuit.cscorelib2.chat.ChatColors;
+import io.github.thebusybiscuit.slimefun4.core.guide.SlimefunGuide;
+import io.github.thebusybiscuit.slimefun4.core.guide.SlimefunGuideImplementation;
+import io.github.thebusybiscuit.slimefun4.implementation.SlimefunPlugin;
+import io.github.thebusybiscuit.slimefun4.implementation.guide.CheatSheetSlimefunGuide;
+
+/**
+ * This is just a helper {@link ItemStack} class for the {@link SlimefunGuide} {@link ItemStack}.
+ * 
+ * @author TheBusyBiscuit
+ * 
+ * @see SlimefunGuide
+ * @see SlimefunGuideImplementation
+ *
+ */
+public class SlimefunGuideItem extends ItemStack {
+
+    public SlimefunGuideItem(SlimefunGuideImplementation implementation, String name) {
+        super(Material.ENCHANTED_BOOK);
+
+        ItemMeta meta = getItemMeta();
+
+        meta.setDisplayName(ChatColors.color(name));
+
+        List<String> lore = new LinkedList<>();
+
+        lore.add(implementation instanceof CheatSheetSlimefunGuide ? "&4&lOnly openable by Admins" : "");
+        lore.add(ChatColors.color("&eRight Click &8\u21E8 &7Browse Items"));
+        lore.add(ChatColors.color("&eShift + Right Click &8\u21E8 &7Open Settings / Credits"));
+
+        meta.setLore(lore);
+        SlimefunPlugin.getItemTextureService().setTexture(meta, "SLIMEFUN_GUIDE");
+
+        setItemMeta(meta);
+    }
+
+}