+++ src/main/java/me/mrCookieSlime/Slimefun/Objects/SlimefunItem/abstractItems/AContainer.java
@@ -109,7 +109,9 @@ public abstract class AContainer extends SlimefunItem implements InventoryBlock,
      * 
      * @return The title of the {@link Inventory} of this {@link AContainer}
      */
-    public abstract String getInventoryTitle();
+    public String getInventoryTitle() {
+        return getItemName();
+    }
 
     /**
      * This method returns the {@link ItemStack} that this {@link AContainer} will
@@ -258,7 +260,7 @@ public abstract class AContainer extends SlimefunItem implements InventoryBlock,
             }
         }
         else {
-            MachineRecipe next = nextRecipe(inv);
+            MachineRecipe next = findNextRecipe(inv);
 
             if (next != null) {
                 processing.put(b, next);
@@ -267,7 +269,7 @@ public abstract class AContainer extends SlimefunItem implements InventoryBlock,
         }
     }
 
-    private MachineRecipe nextRecipe(BlockMenu inv) {
+    protected MachineRecipe findNextRecipe(BlockMenu inv) {
         Map<Integer, ItemStack> inventory = new HashMap<>();
 
         for (int slot : getInputSlots()) {